refactor span decoder
