+++ zipkin-gems/zipkin-query/thrift/zipkinCore.thrift
@@ -1,0 +1,1 @@
+../../../zipkin-thrift/src/main/resources/thrift/zipkinCore.thrift
