+++ zipkin-receiver-kafka/src/main/scala/com/twitter/zipkin/receiver/kafka/KafkaProcessor.scala
@@ -10,6 +10,7 @@ import com.twitter.scrooge.BinaryThriftStructSerializer
 import com.twitter.zipkin.conversions.thrift.{thriftSpanToSpan, spanToThriftSpan}
 import kafka.message.Message
 
+
 object KafkaProcessor {
 
   type KafkaDecoder = Decoder[Option[List[ThriftSpan]]]
@@ -50,20 +51,3 @@ class KafkaProcessor[T](
     }
   }
 }
-
-class SpanDecoder extends KafkaProcessor.KafkaDecoder {
-  val deserializer = new BinaryThriftStructSerializer[ThriftSpan] {
-    def codec = ThriftSpan
-  }
-
-  def toEvent(message: Message): Option[List[ThriftSpan]] = {
-
-    val buffer = message.payload
-    val payload = new Array[Byte](buffer.remaining)
-    buffer.get(payload)
-    val span = deserializer.fromBytes(payload)
-    Some(List(span))
-  }
-  def fromBytes(bytes: Array[Byte]): Option[List[ThriftSpan]] = Some(List(deserializer.fromBytes(bytes)))
-}
-
