+++ zipkin-receiver-kafka/src/main/scala/com/twitter/zipkin/receiver/kafka/KafkaSpanReceiver.scala
@@ -52,5 +52,4 @@ trait KafkaSpanReceiverFactory { self: App =>
       Closable.sequence(service).close(deadline)
     }
   }
-
 }
