+++ zipkin-receiver-kafka/src/main/scala/com/twitter/zipkin/receiver/kafka/KafkaStreamProcessor.scala
@@ -26,5 +26,4 @@ case class KafkaStreamProcessor[T](
         log.error(s"${e.getCause}")
     }
   }
-
 }
